when driver accepted remove from other driver

+++ src/controllers/callTaxi/index.ts
@@ -316,14 +316,16 @@ export const driverUpdateStatus = async (req: Request, res: Response) => {
             return;
         }
 
-        // Delete ride matching from other when once accepted
-        await axios.delete(`${process.env.SOCKET_SERVICE_URL}/v1/api/ride-request-socket/remove/${confirmed?._id}`,
-            {
-                headers: {
-                    Authorization: req.headers.authorization
+        if (status === STATUS.DRIVER_RECEIVED) {
+            // Delete ride matching from other when once accepted
+            await axios.delete(`${process.env.SOCKET_SERVICE_URL}/v1/api/ride-request-socket/remove/${confirmed?._id}`,
+                {
+                    headers: {
+                        Authorization: req.headers.authorization
+                    }
                 }
-            }
-        )
+            )
+        }
 
         res.status(200).json({
             code: messages.SUCCESSFULLY.code,
